requirements.txt
langchain>=0.2.0
langchain-community
langchain-core
cohere
pydantic
wikipedia
pdfplumber
PyPDF2
faiss-cpu
python-dotenv
